rewrite `test_integration_1` tests with fixture
