+++ tests/integration/tab/test_tabs.py
@@ -1,0 +1,127 @@
+import dash
+from dash.dependencies import Input, Output
+import dash_core_components as dcc
+import dash_html_components as html
+
+
+def test_tabs001_in_vertical_mode(dash_dcc):
+    app = dash.Dash(__name__)
+    app.layout = html.Div(
+        [
+            dcc.Tabs(
+                id="tabs",
+                value="tab-3",
+                children=[
+                    dcc.Tab(
+                        label="Tab one",
+                        value="tab-1",
+                        id="tab-1",
+                        children=[html.Div("Tab One Content")],
+                    ),
+                    dcc.Tab(
+                        label="Tab two",
+                        value="tab-2",
+                        id="tab-2",
+                        children=[html.Div("Tab Two Content")],
+                    ),
+                    dcc.Tab(
+                        label="Tab three",
+                        value="tab-3",
+                        id="tab-3",
+                        children=[html.Div("Tab Three Content")],
+                    ),
+                ],
+                vertical=True,
+            ),
+            html.Div(id="tabs-content"),
+        ]
+    )
+
+    dash_dcc.start_server(app)
+    dash_dcc.wait_for_text_to_equal('#tab-3', 'Tab three')
+    dash_dcc.percy_snapshot('Tabs - vertical mode')
+    assert dash_dcc.get_logs() == []
+
+
+def test_tabs002_without_children(dash_dcc):
+    app = dash.Dash(__name__)
+    app.layout = html.Div(
+        [
+            html.H1("Dash Tabs component demo"),
+            dcc.Tabs(
+                id="tabs",
+                value="tab-2",
+                children=[
+                    dcc.Tab(label="Tab one", value="tab-1", id="tab-1"),
+                    dcc.Tab(label="Tab two", value="tab-2", id="tab-2"),
+                ],
+            ),
+            html.Div(id="tabs-content"),
+        ]
+    )
+
+    @app.callback(
+        dash.dependencies.Output("tabs-content", "children"),
+        [dash.dependencies.Input("tabs", "value")],
+    )
+    def render_content(tab):
+        if tab == "tab-1":
+            return html.Div([html.H3("Test content 1")], id="test-tab-1")
+        elif tab == "tab-2":
+            return html.Div([html.H3("Test content 2")], id="test-tab-2")
+
+    dash_dcc.start_server(app)
+    dash_dcc.wait_for_text_to_equal('#tabs-content', 'Test content 2')
+    dash_dcc.percy_snapshot('initial tab - tab 2')
+
+    dash_dcc.wait_for_element('#tab-1').click()
+    dash_dcc.wait_for_text_to_equal('#tabs-content', 'Test content1')
+    assert dash_dcc.get_logs() == []
+
+
+def test_tabs003_without_children_undefined(dash_dcc):
+    app = dash.Dash(__name__)
+    app.layout = html.Div(
+        [
+            html.H1("Dash Tabs component demo"),
+            dcc.Tabs(id="tabs", value="tab-1"),
+            html.Div(id="tabs-content"),
+        ]
+    )
+
+    dash_dcc.start_server(app)
+    dash_dcc.wait_for_element('#tabs-content')
+    dash_dcc.percy_snapshot('Tabs component with children undefined')
+    assert dash_dcc.get_logs() == []
+
+
+def test_tabs004_without_value(dash_dcc):
+    app = dash.Dash(__name__)
+
+    app.layout = html.Div(
+        [
+            html.H1("Dash Tabs component demo"),
+            dcc.Tabs(
+                id="tabs-without-value",
+                children=[
+                    dcc.Tab(label="Tab One", value="tab-1"),
+                    dcc.Tab(label="Tab Two", value="tab-2"),
+                ],
+            ),
+            html.Div(id="tabs-content"),
+        ]
+    )
+
+    @app.callback(
+        Output("tabs-content", "children"), [Input("tabs-without-value", "value")]
+    )
+    def render_content(tab):
+        if tab == "tab-1":
+            return html.H3("Default selected Tab content 1")
+        elif tab == "tab-2":
+            return html.H3("Tab content 2")
+
+    dash_dcc.start_server(app)
+    dash_dcc.wait_for_text_to_equal('#tabs-content', 'Default selected Tab content 1')
+    dash_dcc.percy_snapshot('Tab 1 should be selected by default')
+    assert dash_dcc.get_logs() == []
